==> analyses/survival/01-prepare-survival.Rmd <==
---
title: "DEI Survival Analyses"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Ryan Corbett
date: 2022
params:
  plot_ci: TRUE
---

**Purpose:** 

Prepare ancestry data for survival analyses by adding most recent survival data and coding event-free survival (EFS)

## Usage 

Uses a wrapper function (`survival_analysis`) from utils folder. 

## Setup

#### Packages and functions

Set up directory paths

```{r Set up library}
library(survival)
library(ggpubr)
library(tidyverse)
# Set directories
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
data_dir <- file.path(root_dir, "data")
analysis_dir <- file.path(root_dir, "analyses", "survival")

input_dir <- file.path(analysis_dir, "input")
results_dir <- file.path(analysis_dir, "results")

if (!dir.exists(results_dir)) {
  dir.create(results_dir)
}
```

Declare input file paths and useful variables

```{r Set input file paths}
ancestry_file <- file.path(root_dir, "analyses", "add-histologies", "results", "merged_ancestry_histology_data.tsv")

survival_file <- file.path(data_dir, "survival-histologies-base-2023-09-20.tsv")

output_file <- file.path(results_dir, "merged_ancestry_histology_survival.tsv")
```

## Import the metadata and append updated EFS data from base histologies

```{r Read in metadata}
ancestry <- readr::read_tsv(ancestry_file)

# survival <- read_tsv(survival_file)

ancestry <- ancestry %>%
  # left_join(survival[,c("Kids_First_Participant_ID", "OS_days", "OS_status", "EFS_days", "EFS_event_type")]) %>%
  # distinct(Kids_First_Participant_ID, .keep_all = T) %>%
  dplyr::mutate(EFS_days = as.numeric(EFS_days),
                OS_days = as.numeric(OS_days)) %>%
  dplyr::mutate(OS_years = OS_days/365.25,
         EFS_years = EFS_days/365.25) %>%
  dplyr::mutate(EFS_status = case_when(
    EFS_event_type == "Not Applicable" ~ "NO EVENT",
    !is.na(EFS_event_type) ~ "EVENT",
    TRUE ~ NA_character_
  )) %>%
  dplyr::mutate(EUR_status = case_when(
    predicted_ancestry == "EUR" ~ "EUR", 
    TRUE ~ "non-EUR"
  ))

```


```{r format mol sub gruop}

ancestry <- ancestry %>%
    # The molecular subtypes for several DIPG/DMG samples is incorrectly assigned as an HGG subtype; modify here:
    mutate(molecular_subtype = case_when(
      plot_group == "DIPG or DMG" & grepl("HGG", molecular_subtype) & grepl("TP53", molecular_subtype) ~ "DMG, H3 K28, TP53",
      plot_group == "DIPG or DMG" & grepl("HGG", molecular_subtype) ~ "DMG, H3 K28",
      TRUE ~ molecular_subtype
    )) %>%
    mutate(mol_sub_group = case_when(
    plot_group %in% c("DIPG or DMG", "Other high-grade glioma") & grepl("wildtype", molecular_subtype) ~ "HGG, H3 WT",
    plot_group %in% c("DIPG or DMG", "Other high-grade glioma") & grepl("K28", molecular_subtype) ~ "DMG, H3 K28",
    plot_group %in% c("DIPG or DMG", "Other high-grade glioma") & grepl("G35", molecular_subtype) ~ "DHG, H3 G35",
    plot_group %in% c("DIPG or DMG", "Other high-grade glioma") & grepl("IHG", molecular_subtype) ~ "IHG",
    plot_group %in% c("DIPG or DMG", "Other high-grade glioma") & grepl("IDH", molecular_subtype) ~ "HGG, IDH",
    plot_group %in% c("Low-grade glioma") & grepl("SEGA", molecular_subtype) ~ "SEGA",
    plot_group %in% c("Low-grade glioma") & grepl("V600E", molecular_subtype) ~ "LGG, BRAF V600E",
    plot_group %in% c("Low-grade glioma") & grepl("-BRAF", molecular_subtype) ~ "LGG, BRAF fusion",
    plot_group %in% c("Low-grade glioma") & grepl("wildtype", molecular_subtype) ~ "LGG, WT",
    plot_group %in% c("Low-grade glioma") & grepl("To be classified", molecular_subtype) ~ "LGG, To be classified",
    plot_group %in% c("Low-grade glioma") ~ "LGG, Other alteration",
    plot_group %in% c("Mixed neuronal-glial tumor") & grepl("wildtype", molecular_subtype) ~ "GNG/GNT, WT",
    plot_group %in% c("Mixed neuronal-glial tumor") & grepl("V600E", molecular_subtype) ~ "GNG/GNT, BRAF V600E",
    plot_group %in% c("Mixed neuronal-glial tumor") & grepl("-BRAF", molecular_subtype) ~ "GNG/GNT, BRAF fusion",
    plot_group %in% c("Mixed neuronal-glial tumor") & !grepl("To be classified", molecular_subtype) ~ "GNG/GNT, Other alteration",
    !is.na(molecular_subtype) ~ molecular_subtype
  )) %>%
  mutate(age_at_diagnosis_days = as.numeric(age_at_diagnosis_days)) %>%
  mutate(age_at_diagnosis_years = age_at_diagnosis_days/365.25)

```

```{r save output}
write_tsv(ancestry, output_file)

```

```{r sessionInfo}
sessionInfo()

```
