==> analyses/survival/02-run-survival.Rmd <==
---
title: "PBTA Ancestry Survival Analyses"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Ryan Corbett
date: 2023
---

**Purpose:** 

Runs survival analysis models within tumor histologies and molecular subtypes by predicted ancestry, self-reported race, and self-reported ethnicity for PBTA-ancestry cohort

## Setup

#### Packages and functions

Read in set up script.

```{r Set up library}
library(survival)
library(ggpubr)
library(tidyverse)
library(ComplexHeatmap)
library(circlize)

# Set directories
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
analysis_dir <- file.path(root_dir, "analyses", "survival")
# Source this script, which contains a wrapper function that can conduct the survival analyses, from OpenPBTA
source(file.path(analysis_dir, "util", "survival_models_additive.R"))
source(file.path(analysis_dir, "util", "heatmap_function.R"))
# Magrittr pipe
`%>%` <- dplyr::`%>%`
```

#### Set up files and directories

Set up output directories. 
```{r Set up directories}
input_dir <- file.path(root_dir, "analyses", "survival", "results")
results_dir <- file.path(analysis_dir, "results")
plots_dir <- file.path(analysis_dir, "plots")
```

Make output directories.
```{r Make output directories}
if (!dir.exists(results_dir)) {
  dir.create(results_dir)
}
if (!dir.exists(plots_dir)) {
  dir.create(plots_dir)
}
```

Declare input file paths and useful variables
```{r Set input file paths}
ancestry_file <- file.path(input_dir, "merged_ancestry_histology_survival.tsv")
```


## Import the metadata
```{r Read in metadata}
metadata <- readr::read_tsv(ancestry_file)

metadata <- metadata %>%
  dplyr::mutate(predicted_ancestry = factor(predicted_ancestry,
                                            levels = c("AFR", "AMR", "EAS", "EUR", "SAS")
                                            )) %>%
  dplyr::mutate(extent_of_tumor_resection = case_when(
          extent_of_tumor_resection %in% c('Not Reported', "Unavailable") ~ NA_character_,
          extent_of_tumor_resection == "Not Applicable" ~ "No resection",
          TRUE ~ extent_of_tumor_resection
        )) %>%
  dplyr::mutate(extent_of_tumor_resection = fct_relevel(extent_of_tumor_resection,
                                                        c("Gross/Near total resection",
                                                          "Partial resection", "Biopsy only")))
  
```

Define plot groups for which survival analyses should be run
```{r define histologies}
# Define groups for survival analyses
groups <- c("Atypical Teratoid Rhabdoid Tumor", "Choroid plexus tumor",
            "Craniopharyngioma", "DIPG or DMG", 
            "Other high-grade glioma", "Ependymoma",
            "Germ cell tumor",
            "Mixed neuronal-glial tumor", "Low-grade glioma",
            "Medulloblastoma", "Meningioma",
            "Mesenchymal tumor", "Neurofibroma plexiform", 
            "Oligodendroglioma", "Schwannoma"
)

# create names vector for output files
plot_names <- c("ATRT", "CPT", "CRANIO", "DMG", 
                "HGG", "EPN", "GERM", "GNG", 
                "LGG", "MB", "MNG", "MES", 
                "NFP", "OLIGO", "SWN")
names(plot_names) <- groups
```

Loop through groups and create Kaplan Meier survival and Cox Proportional Hazard models for OS and EFS
```{r histology survival}
variable <- c("predicted_ancestry", "EUR_status")

for (var in variable){

  for (group in groups){
    
    results_dir <- file.path(analysis_dir, "results", plot_names[group])
    
    if (!dir.exists(results_dir)) {
      dir.create(results_dir)
      
      }
    
    # Define output files
    km_os_model_file <- file.path(results_dir, 
                                      glue::glue("logrank_{plot_names[group]}_OS_{var}.RDS"))

    coxph_os_model_file <- ifelse(group %in% c("Low-grade glioma", "Atypical Teratoid Rhabdoid Tumor",
                                               "Medulloblastoma", "Other high-grade glioma", "Schwannoma"),
                               file.path(results_dir,
                                         glue::glue("cox_{plot_names[group]}_OS_additive_terms_subtype_resection_{var}.RDS")),
                               file.path(results_dir,
                                         glue::glue("cox_{plot_names[group]}_OS_additive_terms_subtype_{var}.RDS"))
    )
    
    km_efs_model_file <- file.path(results_dir, 
                                      glue::glue("logrank_{plot_names[group]}_EFS_{var}.RDS"))

    coxph_efs_model_file <- ifelse(group %in% c("Low-grade glioma", "Atypical Teratoid Rhabdoid Tumor",
                                               "Medulloblastoma", "Other high-grade glioma", "Schwannoma"),
                               file.path(results_dir,
                                         glue::glue("cox_{plot_names[group]}_EFS_additive_terms_subtype_resection_{var}.RDS")),
                               file.path(results_dir,
                                         glue::glue("cox_{plot_names[group]}_EFS_additive_terms_subtype_{var}.RDS"))
    )
    
    subtype_plot_pdf <- file.path(plots_dir, 
                                  glue::glue("{plot_names[group]}_subtype_by_{var}.pdf"))
    
    # Subset histology
    group_hist <- metadata %>%
      filter(grepl(group, plot_group)) %>%
      dplyr::mutate(mol_sub_group = factor(mol_sub_group))
    
    if (group == "Other high-grade glioma"){
      group_hist <- group_hist %>%
       mutate(molecular_subtype = fct_relevel(molecular_subtype, c("HGG, H3 wildtype",
                                                           "HGG, H3 wildtype, TP53",
                                                       "DHG, H3 G35",
                                                       "DHG, H3 G35, TP53",
                                                       "HGG, IDH, TP53",
                                                       "HGG, PXA", "IHG, ALK-altered",
                                                       "IHG, NTRK-altered",
                                                       "IHG, NTRK-altered, TP53",
                                                       "IHG, To be classified")))
    } 
    
    if (group == "DIPG or DMG"){
      group_hist <- group_hist %>%
       mutate(molecular_subtype = fct_relevel(molecular_subtype, c("DMG, H3 K28",
                                                           "DMG, H3 K28, TP53")))
    } 
    
    if (group == "Ependymoma") {
      group_hist <- group_hist %>%
        dplyr::mutate(mol_sub_group = fct_relevel(mol_sub_group,
                      c("EPN, ST ZFTA", "EPN, ST YAP1", "EPN, PF A", 
                        "EPN, PF B", "EPN, MPE", "EPN, SP", "EPN, SP-MYCN",
                        "EPN, To be classified")))
        
    } 
    
    if (group == "Low-grade glioma") {
      group_hist <- group_hist %>%
        dplyr::mutate(mol_sub_group = fct_relevel(mol_sub_group,
                                                  c("LGG, WT", "LGG, BRAF V600E",
                                                    "LGG, BRAF fusion",
                                                    "LGG, Other alteration",
                                                    "SEGA",
                                                    "LGG, To be classified")))
    }
    
    if (group == "Medulloblastoma") {
      group_hist <- group_hist %>%
        dplyr::mutate(mol_sub_group = fct_relevel(mol_sub_group,
                                                  c("MB, WNT", "MB, Group3",
                                                    "MB, Group 4", "MB, SHH")))
    }
    
    if (group == "Mesenchymal tumor") {
      group_hist <- group_hist %>%
        dplyr::mutate(cancer_group = fct_relevel(cancer_group,
                                                  c("Sarcoma", "Chordoma", "Chondromyxoid fibroma",
                                                    "Ewing sarcoma", "Hemangioblastoma",
                                                    "Inflammatory Myofibroblastic Tumor",
                                                    "Mesenchymal tumor")))
    }
    
    if (group == "Mixed neuronal-glial tumor") {
      group_hist <- group_hist %>%
        dplyr::mutate(mol_sub_group = fct_relevel(mol_sub_group,
                                                  c("GNG/GNT, WT", "GNG/GNT, BRAF V600E",
                                                    "GNG/GNT, BRAF fusion",
                                                    "GNG/GNT, Other alteration",
                                                    "GNG, To be classified",
                                                    "GNT, To be classified")))
    }

    # create heatmap of subtype by predicted ancestry
    if (var == "predicted_ancestry" & (length(table(group_hist$mol_sub_group)) >= 2 | group == "DIPG or DMG")) {

      if (group %in% c("Other high-grade glioma", "DIPG or DMG")){
        
        pdf(subtype_plot_pdf,
          height = length(2 + unique(group_hist$predicted_ancestry[!is.na(group_hist$molecular_subtype) & !grepl("classified", group_hist$molecular_subtype)]))/1.25, 
          width = 6)
      
        subgroup <- "molecular_subtype"
        subtypes <- levels(group_hist$molecular_subtype)[!grepl("classified", levels(group_hist$molecular_subtype))]
        
        group_ht <- plot_enr(group_hist[!is.na(group_hist$molecular_subtype) & !grepl("classified", group_hist$molecular_subtype),], 
                     var1 = subgroup, 
                     var2 = "predicted_ancestry",
                     var1_names = subtypes,
                     var2_names = sort(unique(group_hist$predicted_ancestry[!is.na(group_hist$molecular_subtype) & !grepl("classified", group_hist$molecular_subtype)])),
                     padjust = TRUE)

      } else {
        
        pdf(subtype_plot_pdf,
          height = length(2 + unique(group_hist$predicted_ancestry[!is.na(group_hist$mol_sub_group) & !grepl("classified", group_hist$mol_sub_group)]))/1.5, 
          width = 6)
        
        subgroup <- "mol_sub_group"
        subtypes <- levels(group_hist$mol_sub_group)[!grepl("classified", levels(group_hist$mol_sub_group))]
        
        group_ht <- plot_enr(group_hist[!is.na(group_hist$mol_sub_group) & !grepl("classified", group_hist$mol_sub_group),], 
                     var1 = subgroup, 
                     var2 = "predicted_ancestry",
                     var1_names = subtypes,
                     var2_names = sort(unique(group_hist$predicted_ancestry[!is.na(group_hist$mol_sub_group) & !grepl("classified", group_hist$mol_sub_group)])),
                     padjust = TRUE)
        
      }

      draw(group_ht)
      
      invisible(dev.off())
      
    }

    # Run kaplain-meier survival models
    kap_os <- survival_analysis(
    metadata  = group_hist[!grepl("To be classified", group_hist$mol_sub_group),],
    ind_var = var,
    test = "kap.meier",
    metadata_sample_col = "Kids_First_Biospecimen_ID",
    days_col = "OS_days",
    status_col = "OS_status"
    )
  
    readr::write_rds(kap_os, km_os_model_file)
    
    kap_efs <- survival_analysis(
    metadata  = group_hist[!grepl("To be classified", group_hist$mol_sub_group),],
    ind_var = var,
    test = "kap.meier",
    metadata_sample_col = "Kids_First_Biospecimen_ID",
    days_col = "EFS_days",
    status_col = "EFS_status"
    )
  
    readr::write_rds(kap_efs, km_efs_model_file)
    
    group_hist <- group_hist %>%
      mutate(predicted_ancestry = factor(predicted_ancestry,
                                            c("EUR", "AFR", "AMR", "EAS", "SAS")))
  
    # Run cox proportional hazards model
    add_model_os <- fit_save_model(group_hist[!grepl("To be classified", group_hist$mol_sub_group) & !grepl("To be classified", group_hist$molecular_subtype),],
                                    terms = ifelse(
                                        group %in% c("Low-grade glioma", "Atypical Teratoid Rhabdoid Tumor", "Medulloblastoma"),
                                        glue::glue("extent_of_tumor_resection+mol_sub_group+age_at_diagnosis_years+{var}"),
                                        ifelse(
                                          group == "Other high-grade glioma",
                                          glue::glue("extent_of_tumor_resection+molecular_subtype+age_at_diagnosis_years+{var}"),
                                          ifelse(
                                            group == "DIPG or DMG", 
                                            glue::glue("molecular_subtype+age_at_diagnosis_years+{var}"),
                                            ifelse(
                                              group %in% c("Ependymoma", "Medulloblastoma", "Mixed neuronal-glial tumor"),
                                              glue::glue("mol_sub_group+age_at_diagnosis_years+{var}"),
                                              ifelse(
                                                group %in% c("Germ cell tumor", "Mesenchymal tumor"),
                                                glue::glue("cancer_group+age_at_diagnosis_years+{var}"),
                                                ifelse(group == "Schwannoma",
                                                       glue::glue("extent_of_tumor_resection+age_at_diagnosis_years+{var}"),
                                                       glue::glue("age_at_diagnosis_years+{var}")
                                                        )))))),
               coxph_os_model_file,
               "multivariate",
               years_col = "OS_years", status_col = "OS_status"
              )
    
    add_model_efs <- fit_save_model(group_hist[!grepl("To be classified", group_hist$mol_sub_group) & !grepl("To be classified", group_hist$molecular_subtype),],
                                    terms = ifelse(
                                        group %in% c("Low-grade glioma", "Atypical Teratoid Rhabdoid Tumor", "Medulloblastoma"),
                                        glue::glue("extent_of_tumor_resection+mol_sub_group+age_at_diagnosis_years+{var}"),
                                        ifelse(
                                          group == "Other high-grade glioma",
                                          glue::glue("extent_of_tumor_resection+molecular_subtype+age_at_diagnosis_years+{var}"),
                                          ifelse(
                                            group == "DIPG or DMG", 
                                            glue::glue("molecular_subtype+age_at_diagnosis_years+{var}"),
                                            ifelse(
                                              group %in% c("Ependymoma", "Medulloblastoma", "Mixed neuronal-glial tumor"),
                                              glue::glue("mol_sub_group+age_at_diagnosis_years+{var}"),
                                              ifelse(
                                                group %in% c("Germ cell tumor", "Mesenchymal tumor"),
                                                glue::glue("cancer_group+age_at_diagnosis_years+{var}"),
                                                ifelse(group == "Schwannoma",
                                                       glue::glue("extent_of_tumor_resection+age_at_diagnosis_years+{var}"),
                                                       glue::glue("age_at_diagnosis_years+{var}")
                                                        )))))),
               coxph_efs_model_file,
               "multivariate",
               years_col = "EFS_years", status_col = "EFS_status"
              )
    
  }
}

```

Loop through molecular subtypes to generate subtype-specific survival models by predicted ancestry, reported race, and reported ancestry
```{r subtype survival}

# define subtype df including subtypes with n>=20
subtype_df <- metadata %>%
  count(mol_sub_group, plot_group) %>%
  filter(n >=20 & !grepl("classified", mol_sub_group) & !is.na(mol_sub_group)) %>%
  dplyr::mutate(hist = unlist(lapply(strsplit(mol_sub_group, ", "), function(x) x[1])),
                subtype = case_when(
                  grepl(",", mol_sub_group) ~ unlist(lapply(strsplit(mol_sub_group, ", "), function(x) x[2])),
                  TRUE ~ NA_character_)) %>%
  dplyr::mutate(subtype = str_replace_all(subtype, " ", "-")) %>%
  dplyr::mutate(hist = case_when(
    hist == "GNG/GNT" ~ "GNG",
    TRUE ~ hist
  ))

variable <- c("predicted_ancestry", "EUR_status")

for (var in variable){
  
  for (i in 1:nrow(subtype_df)) {
    
    results_dir <- file.path(analysis_dir, "results", subtype_df$hist[i])
    
    if (!dir.exists(results_dir)) {
      dir.create(results_dir)
      
      }
    
    km_os_model_file <- file.path(results_dir, 
                                      glue::glue("logrank_{subtype_df$hist[i]}_{subtype_df$subtype[i]}_OS_{var}.RDS"))
  
    coxph_os_model_file <- ifelse(subtype_df$hist[i] %in% c("LGG", "HGG", "MB"),
                               file.path(results_dir,
                                        glue::glue("cox_{subtype_df$hist[i]}_{subtype_df$subtype[i]}_OS_additive_terms_subtype_resection_{var}.RDS")),
                               file.path(results_dir,
                                         glue::glue("cox_{subtype_df$hist[i]}_{subtype_df$subtype[i]}_OS_additive_terms_subtype_{var}.RDS"))
    )
  
      km_efs_model_file <- file.path(results_dir, 
                                      glue::glue("logrank_{subtype_df$hist[i]}_{subtype_df$subtype[i]}_EFS_{var}.RDS"))
  
    coxph_efs_model_file <- ifelse(subtype_df$hist[i] %in% c("LGG", "HGG", "MB"),
                               file.path(results_dir,
                                        glue::glue("cox_{subtype_df$hist[i]}_{subtype_df$subtype[i]}_EFS_additive_terms_subtype_resection_{var}.RDS")),
                               file.path(results_dir,
                                         glue::glue("cox_{subtype_df$hist[i]}_{subtype_df$subtype[i]}_EFS_additive_terms_subtype_{var}.RDS"))
    )
    
    # subset data for subtype
    subtype_hist <- metadata %>%
      filter(plot_group == subtype_df$plot_group[i] & grepl(subtype_df$mol_sub_group[i], mol_sub_group))
      
    # Run Kaplan-Meier survival models
    kap_os <- survival_analysis(
    metadata  = subtype_hist,
    ind_var = var,
    test = "kap.meier",
    metadata_sample_col = "Kids_First_Biospecimen_ID",
    days_col = "OS_days",
    status_col = "OS_status"
    )
    
    readr::write_rds(kap_os, km_os_model_file)
    
    kap_efs <- survival_analysis(
    metadata  = subtype_hist,
    ind_var = var,
    test = "kap.meier",
    metadata_sample_col = "Kids_First_Biospecimen_ID",
    days_col = "EFS_days",
    status_col = "EFS_status"
    )
    
    readr::write_rds(kap_efs, km_efs_model_file)
    
    subtype_hist <- subtype_hist %>%
      mutate(predicted_ancestry = factor(predicted_ancestry,
                                                c("EUR", "AFR", "AMR", "EAS", "SAS")))

    add_model_os <- fit_save_model(subtype_hist,
                                  terms = ifelse(subtype_df$plot_group[i] %in% c("Low-grade glioma", "Medulloblastoma"),
                                                 glue::glue("extent_of_tumor_resection+age_at_diagnosis_years+{var}"),
                                                 ifelse(subtype_df$plot_group[i] == "Other high-grade glioma",
                                                        glue::glue("extent_of_tumor_resection+molecular_subtype+age_at_diagnosis_years+{var}"),
                                                        glue::glue("age_at_diagnosis_years+{var}"))),
                                   coxph_os_model_file,
                                   "multivariate",
                                   years_col = "OS_years", status_col = "OS_status"
                                  )
    
    add_model_efs <- fit_save_model(subtype_hist,
                                  terms = ifelse(subtype_df$plot_group[i] %in% c("Low-grade glioma", "Medulloblastoma"),
                                                 glue::glue("extent_of_tumor_resection+age_at_diagnosis_years+{var}"),
                                                 ifelse(subtype_df$plot_group[i] == "Other high-grade glioma",
                                                        glue::glue("extent_of_tumor_resection+molecular_subtype+age_at_diagnosis_years+{var}"),
                                                        glue::glue("age_at_diagnosis_years+{var}"))),
                                     coxph_efs_model_file,
                                     "multivariate",
                                     years_col = "EFS_years", status_col = "EFS_status"
                                    )
    
  }
    
}

```

Print session info
```{r sessionInfo}
sessionInfo()

```
